refactor(instance): split workflows.rs god object into modules
