+++ scripts/.god-object-baseline.txt
@@ -4,7 +4,6 @@
 1106 ui/src/routes/instances.$id.tsx
 1095 ai-agent-sandbox-blueprint-lib/src/workflows.rs
 1065 sandbox-runtime/src/session_auth.rs
-948 ai-agent-instance-blueprint-lib/src/workflows.rs
 922 ui/src/routes/workflows._index.tsx
 838 sandbox-runtime/src/reaper.rs
 744 sandbox-runtime/src/scoped_session_auth.rs
